fix: rename config namespace for auto tracking
