+++ packages/analytics-types/src/config/index.ts
@@ -1,4 +1,4 @@
 export { Config, Options } from './core';
-export { BrowserConfig, DefaultTrackingOptions, TrackingOptions, AttributionOptions, BrowserOptions } from './browser';
+export { BrowserConfig, AutoTrackingOptions, TrackingOptions, AttributionOptions, BrowserOptions } from './browser';
 export { NodeConfig, NodeOptions } from './node';
 export { ReactNativeConfig, ReactNativeTrackingOptions, ReactNativeOptions } from './react-native';
